remove importHook and relocate code
